Remove trace_target env var in favor of .next/trace (#31697)
