+++ packages/next/trace/shared.ts
@@ -1,11 +1,4 @@
 // eslint typescript has a bug with TS enums
-/* eslint-disable no-shadow */
-export enum TARGET {
-  CONSOLE = 'CONSOLE',
-  ZIPKIN = 'ZIPKIN',
-  JAEGER = 'JAEGER',
-  TELEMETRY = 'TELEMETRY',
-}
 
 export type SpanId = string
